Ampere: Test Automation: Add test for IPMI chassis policy always-off
This patch creates the test code for IPMI chassis policy always-off
command.
The test case is PASSED if the chassis power status is off after BMC
is reset (simulate the case of BMC power is restored)

This patch passed the rflint verify test code tool.
0 error + warning after using rflint tool to verify the test code.

Redmine ticket #2224
(This patch is already reviewed)

Change-Id: Id63df662cf93f17c1cc945244831f5ebca85e011
Signed-off-by: Hoang Nguyen <hnguyen@amperecomputing.com>