Ampere: redfish: Add test case for delete non-existing IPv4 Address.
Implmeneted by appending to the current IPv4Addresses record with a null
entry.

Also, add new keyword for handling Patch request since:
* rest_client.robot doesn't have a keyword for it.
* We can't sending PATCH request to modify the resource with Auth
Cookie. It seems that bmcweb doesn't have an implementation for it or
it doesn't allow, so the "Redfish Patch Request" keyword use Basic Auth
instead as a workaround.

Implements #2331

Change-Id: I3c203457a5fd4b87e5c6eb4856e7cbcab6afed97
Signed-off-by: Hy Che <hche@amperecomputing.com>